+++ src/chains/RobinHood/TakerSubmitted.sol
@@ -25,8 +25,7 @@ contract RobinHoodSettler is Settler, RobinHoodMixin {
         if (super._dispatchVIP(action, data)) {
             return true;
         } else if ((action == uint32(ISettlerActions.UNISWAPV4_VIP.selector))
-            .or(action == uint32(ISettlerActions.EKUBOV3_VIP.selector))
-            .or(action == uint32(ISettlerActions.PANCAKE_INFINITY_VIP.selector))) {
+            .or(action == uint32(ISettlerActions.EKUBOV3_VIP.selector))) {
             (
                 address recipient,
                 ISignatureTransfer.PermitTransferFrom memory permit,
@@ -42,11 +41,28 @@ contract RobinHoodSettler is Settler, RobinHoodMixin {
 
             if (action == uint32(ISettlerActions.UNISWAPV4_VIP.selector)) {
                 sellToUniswapV4VIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
-            } else if (action == uint32(ISettlerActions.EKUBOV3_VIP.selector)) {
+            } else { // if (action == uint32(ISettlerActions.EKUBOV3_VIP.selector))
                 sellToEkuboV3VIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
-            } else { // if (action == uint32(ISettlerActions.PANCAKE_INFINITY_VIP.selector))
-                sellToPancakeInfinityVIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
             }
+        } else if (action == uint32(ISettlerActions.PANCAKE_INFINITY_VIP.selector)) {
+            (
+                address recipient,
+                ISignatureTransfer.PermitTransferFrom memory permit,
+                uint8 forkId,
+                bool feeOnTransfer,
+                uint256 hashMul,
+                uint256 hashMod,
+                bytes memory fills,
+                bytes memory sig,
+                uint256 amountOutMin
+            ) = abi.decode(
+                data,
+                (address, ISignatureTransfer.PermitTransferFrom, uint8, bool, uint256, uint256, bytes, bytes, uint256)
+            );
+
+            sellToPancakeInfinityVIP(
+                recipient, forkId, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin
+            );
         } else {
             return false;
         }